+++ lib/src/response.dart
@@ -1,7 +1,10 @@
 import 'package:http/http.dart' as http;
+import 'package:postgrest_crud/postgrest_crud.dart';
 
 /// Provides models and offset, limit, count from 'Content-Range' response header.
 class Response<T> {
+  final List<JsonObject> json;
+
   /// List of models from response.
   final List<T> models;
 
@@ -13,7 +16,7 @@ class Response<T> {
   int _offset = 0;
 
   /// Create a response object with models and paging related information.
-  Response({required this.response, required this.models}) {
+  Response({required this.response, required this.json, required this.models}) {
     String? contentRange;
     if (response.headers.containsKey('Content-Range')) {
       contentRange = response.headers['Content-Range']!;